Fixup unit tests
