+++ tests/system_tests/hyperion/external_interaction/conftest.py
@@ -361,7 +361,9 @@ def fgs_composite_for_fake_zocalo(
     config_client,
     hyperion_flyscan_xrc_composite: HyperionGridDetectThenXRayCentreComposite,
     zocalo_for_fake_zocalo: ZocaloResults,
+    oav_for_system_test: OAV,
 ) -> HyperionGridDetectThenXRayCentreComposite:
+    hyperion_flyscan_xrc_composite.oav = oav_for_system_test
     set_mock_value(
         hyperion_flyscan_xrc_composite.aperture_scatterguard.aperture.z.user_setpoint, 2
     )
