+++ tests/unit_tests/conftest.py
@@ -39,7 +39,7 @@ from dodal.devices.fast_grid_scan import (
 from dodal.devices.flux import Flux
 from dodal.devices.hutch_shutter import ShutterState
 from dodal.devices.oav.oav_detector import OAV
-from dodal.devices.oav.oav_parameters import OAVConfigBeamCentre, OAVParameters
+from dodal.devices.oav.oav_parameters import OAVParameters
 from dodal.devices.oav.pin_image_recognition import PinTipDetection
 from dodal.devices.robot import BartRobot
 from dodal.devices.s4_slit_gaps import S4SlitGaps
@@ -417,7 +417,6 @@ async def fake_fgs_composite(
         eiger=i03.eiger.build(mock=True),
         gonio=smargon,
         synchrotron=synchrotron,
-        zocalo=zocalo,
     )
 
     fake_composite.eiger.stage = MagicMock(side_effect=lambda: completed_status())
@@ -672,24 +671,6 @@ def oav_parameters_for_rotation(test_config_files) -> OAVParameters:
     return OAVParameters(
         ConfigClient(""), oav_config_json=test_config_files["oav_config_json"]
     )
-
-
-@pytest.fixture
-def oav(test_config_files):
-    parameters = OAVConfigBeamCentre(
-        test_config_files["zoom_params_file"],
-        test_config_files["display_config"],
-        ConfigClient(""),
-    )
-    oav = i03.oav.build(mock=True, connect_immediately=True, params=parameters)
-
-    set_mock_value(oav.zoom_controller.level, "5.0x")
-    set_mock_value(oav.grid_snapshot.x_size, 1024)
-    set_mock_value(oav.grid_snapshot.y_size, 768)
-
-    oav.snapshot.trigger = MagicMock(side_effect=lambda: completed_status())
-    oav.grid_snapshot.trigger = MagicMock(side_effect=lambda: completed_status())
-    yield oav
 
 
 @pytest.fixture()
